Remove obsolete function
